Relocate legacy and pks handler files to their respective sub modules
